refactor scala
