+++ zoo/src/main/scala/com/intel/analytics/zoo/pipeline/api/net/python/PythonZooNet.scala
@@ -15,19 +15,17 @@
  */
 package com.intel.analytics.zoo.pipeline.api.net.python
 
-import java.nio.{ByteOrder, FloatBuffer}
+import java.nio.FloatBuffer
 import java.util.concurrent.{CopyOnWriteArrayList, TimeUnit}
 import java.util.{ArrayList, List => JList}
 
 import com.intel.analytics.bigdl.Module
-import com.intel.analytics.bigdl.dataset.{Sample => JSample}
 import com.intel.analytics.bigdl.nn.abstractnn.{AbstractModule, Activity}
 import com.intel.analytics.bigdl.nn.keras.KerasLayer
 import com.intel.analytics.bigdl.optim._
 import com.intel.analytics.bigdl.python.api.{JTensor, Sample}
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
-import com.intel.analytics.zoo.TFPark._
 import com.intel.analytics.zoo.common.PythonZoo
 import com.intel.analytics.zoo.pipeline.api.Net
 import com.intel.analytics.zoo.pipeline.api.net._
